added access change service

+++ server/service/documentService.js
@@ -10,31 +10,36 @@ const CODE_DEFAULT = 'public class Example {' +
     '\n\t}' +
     '\n}';
 
-const createNewDocument = async (newDocument, userId) => {
+const createNewDocument = async (newDocument, userId, userEmail) => {
     newDocument.createdBy = userId;
-    newDocument.access.set(userId, 'edit');
+    newDocument.access.set(userId, {
+        email: userEmail,
+        access: 'edit'
+    });
     newDocument.lastEditedBy = userId;
     if (newDocument.type === 'code') {
         newDocument.content = CODE_DEFAULT
     } else if (newDocument.type === 'markdown') {
         newDocument.content = MD_DEFAULT
     }
-    await newDocument.save()
     const user = await User.findById(userId)
+    if (!user) throw new Error("File not found.")
     user.files.set(newDocument._id.toString(), newDocument.created)
+    await newDocument.save()
     await user.save()
     return newDocument
 }
 
 const getDocumentsByUser = async userId => {
     const user = await User.findById(userId)
+    if (!user) throw new Error("File not found.")
     let keys = [...user.files.keys()]
     const docs = await Document.find({
             '_id': {
                 $in: keys
             }
         },
-        'filename lastEdited type')
+        'filename lastEdited type createdBy')
     let docInfo = []
     docs.forEach(doc => {
         docInfo.push({
@@ -42,7 +47,8 @@ const getDocumentsByUser = async userId => {
             filename: doc.filename,
             lastEdited: doc.lastEdited,
             type: doc.type,
-            lastOpened: user.files.get(doc._id)
+            createdBy: doc.createdBy,
+            lastOpened: user.files.get(doc._id.toString())
         })
     })
     return docInfo
@@ -50,30 +56,30 @@ const getDocumentsByUser = async userId => {
 
 const getDocumentById = async (docId, userId) => {
     const doc = await Document.findById(docId)
+    if (!doc) throw new Error("File not found.")
     if (doc.accessType.startsWith('public')) {
         return doc
     } else {
         const access = doc.access.get(userId)
         if (access) {
-            doc.access = access
-            try {
-                const user = await User.findById(userId)
-                user.files.set(doc._id.toString(), new Date())
-                await user.save()
-                return doc
-            } catch (e) {
-                throw "User nor found."
+            const user = await User.findById(userId)
+            if (!user) {
+                throw new Error("User nor found.")
             }
+            user.files.set(doc._id.toString(), new Date())
+            await user.save()
+            return doc
         } else {
-            throw "no_access"
+            throw new Error("no_access")
         }
     }
 }
 
 const updateDocument = async (updateDoc, userId) => {
     const doc = await Document.findById(updateDoc.id)
+    if (!doc) throw new Error("File not found.")
     const access = doc.access.get(userId)
-    if (doc.accessType === 'public-edit' || access === 'edit') {
+    if (doc.accessType === 'public-edit' || access.access === 'edit') {
         doc.content = updateDoc.content
         doc.filename = updateDoc.filename
         doc.lastEdited = Date.now()
@@ -84,16 +90,13 @@ const updateDocument = async (updateDoc, userId) => {
             lastEditedBy: doc.lastEditedBy
         })
     } else {
-        throw "no_access"
+        throw new Error("no_access")
     }
-}
-
-const shareFile = () => {
-
 }
 
 const deleteDocument = async (docId, userId) => {
     const user = await User.findById(userId)
+    if (!user) throw new Error("User not found.")
     if (user.files.get(docId)) {
         user.files.delete(docId)
         await user.save()
@@ -103,12 +106,13 @@ const deleteDocument = async (docId, userId) => {
             await Document.deleteOne({_id: docId})
         }
     } else {
-        throw "File is not in user's list."
+        throw new Error("File is not in user's list.")
     }
 }
 
 const makeCopy = async (docId, userId) => {
     const doc = await Document.findById(docId)
+    if (!doc) throw new Error("File not found.")
     doc._id = mongoose.Types.ObjectId()
     doc.isNew = true
     doc.created = new Date()
@@ -118,12 +122,68 @@ const makeCopy = async (docId, userId) => {
     doc.lastEditedBy = userId
     doc.accessType = 'controlled'
     doc.access = new Map()
-    doc.access.set(userId, 'edit')
-    await doc.save()
+
     const user = await User.findById(userId)
+    if (!user) throw new Error("User not found.")
+    doc.access.set(userId, {access: 'edit', email:user.email})
     user.files.set(doc._id.toString(), new Date())
+    await doc.save()
     await user.save()
     return doc._id
+}
+
+const changeAccessType = async (newAccessType, docId, userId) => {
+    const doc = await Document.findById(docId)
+    if (!doc) throw new Error("File not found.")
+    if (doc.createdBy !== userId) {
+        throw new Error('You have no access to change access type.')
+    }
+
+    doc.accessType = newAccessType
+    await doc.save()
+}
+
+const addAccess = async (addUserEmail, docId, requestUserId, access) => {
+    const doc = await Document.findById(docId)
+    if (requestUserId !== doc.createdBy) {
+        throw new Error('You could not modify the access list of this file.')
+    }
+
+    if (!doc) {
+        throw new Error("File not found.")
+    }
+
+    addUserEmail = addUserEmail.toLowerCase()
+    let addUser = await User.findOne({email: addUserEmail})
+    if (!addUser) {
+        throw new Error("User not found, please check the email address you entered.")
+    }
+
+    if (doc.createdBy === addUser._id.toString()) {
+        throw new Error("Can not change access of the creator.")
+    }
+
+    doc.access.set(addUser._id.toString(), {email: addUser.email, access: access})
+    await doc.save()
+}
+
+const removeAccess = async (docId, removeUserId, requestUserId) => {
+    const doc = await Document.findById(docId)
+
+    if (!doc) {
+        throw new Error("File not found.")
+    }
+
+    if (requestUserId !== doc.createdBy) {
+        throw new Error('You could not modify access list to this file.')
+    }
+
+    if (doc.createdBy === removeUserId.id) {
+        throw new Error("Can not change access of the creator.")
+    }
+
+    doc.access.delete(removeUserId.id)
+    await doc.save()
 }
 
 module.exports = {
@@ -132,5 +192,8 @@ module.exports = {
     getDocumentById,
     updateDocument,
     deleteDocument,
-    makeCopy
+    makeCopy,
+    changeAccessType,
+    addAccess,
+    removeAccess
 }
